Implement ability to favorite items
There is now a "Favorite" button beside each city in the saved list. Pressing the button will toggle whether the item is favorited or not. The button is a placeholder and will be updated
for style/functionality reasons in the future. Some helper functions in the Home component have been modified/created to further modularize the code.