Implement storing weather data in local storage when state changes and fix memory leak
When new weather data is stored in state, the state value is then stored in local storage. Now that the Home component unmounts when the user
navigates to a CityInfo page, it is necessary to make sure no changes to Home component state occur while the Home component is unmounted.
Since asynchronous calls were being made in a useEffect hook of the Home component, a cleanup function and logic has been implemented to not make changes
to state while the component is unmounted.