Fix clicking icons causing City Info page to open
The onClick event for the delete and favorite icons used to propogate to the parent onClick event listener, which would cause
the City Info page to open for that city. New onClick functions have been applied to the favorite and delete icons which stop the
event propogation to the parent.